Broke out device tree configs into discrete components. I'm not sure I like this approach, but for now it works.
Feels like these should be per-image configurations -- but I'm not sure how to approach that yet. So for now, this works.

ADS1015 ADC Configuration on I2C is setup for the pidroponic image.
I'll handle this as well in the pipowermon later. As is, this is pretty awesome as it gets the IIO driver setup, so we're not relying upon hwmon/sensors any longer.